% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automaticflux.R
\name{automaticflux}
\alias{automaticflux}
\title{Automatic flux calculation}
\usage{
automaticflux(
  dataframe,
  myauxfile,
  shoulder,
  gastype,
  fluxSeparation,
  force.separation,
  displayPlots,
  method
)
}
\arguments{
\item{dataframe}{blabla}

\item{myauxfile}{blabla}

\item{shoulder}{blabla}

\item{gastype}{blabla}

\item{fluxSeparation}{blabla}

\item{force.separation}{blabla}

\item{displayPlots}{blabla}

\item{method}{blabla}
}
\value{
blabla
}
\description{
Automatic flux calculation
}
\examples{
blabla

}
