% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bubbles.R
\name{find_bubbles}
\alias{find_bubbles}
\title{Automatic bubble detection}
\usage{
find_bubbles(time, conc, window.size)
}
\arguments{
\item{time}{vector; elapsed time over a given incubation.}

\item{conc}{vector; gas measurements of same length as \code{time}.}

\item{window.size}{integer; the size of the moving window used to calculate
variance.}
}
\value{
a data.frame with chunks of elevated variance likely due to ebullition events.
}
\description{
Local variance of gas measurements within a moving window is calculated and
allows for the identification of probable ebullition events.
This function is mainly used by \code{flux.separator()}
}
\examples{
bubbles <- find_bubbles(time = mydata$Etime,
                        conc = mydata$CH4dry_ppb, window.size = 10)

}
\seealso{
See also the function \code{flux.separator()} and
\code{plot.incubations()} for more information about usage.
}
