% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dCdt_density.R
\name{get_dCdt_density}
\alias{get_dCdt_density}
\title{Title}
\usage{
get_dCdt_density(dataframe, gastype)
}
\arguments{
\item{dataframe}{blabla}

\item{gastype}{blabla}
}
\value{
blabla
}
\description{
Title
}
\examples{
blabla


}
