% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_save_pdf.R
\name{gg_save_pdf}
\alias{gg_save_pdf}
\title{Title}
\usage{
gg_save_pdf(list, path, filename)
}
\arguments{
\item{list}{blabla}

\item{path}{blabla}

\item{filename}{blabla}
}
\value{
blabla
}
\description{
Title
}
\examples{
blabla


}
