% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_auxfile_with_data.loop.R
\name{join_auxfile_with_data}
\alias{join_auxfile_with_data}
\title{Title}
\usage{
join_auxfile_with_data(flux.unique)
}
\arguments{
\item{flux.unique}{blabla}
}
\value{
blabla
}
\description{
Title
}
\examples{
blabla

}
