% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_incubations.R
\name{plot_incubations}
\alias{plot_incubations}
\title{A simple plot to provide an overview of gas measurements}
\usage{
plot_incubations(dataframe)
}
\arguments{
\item{dataframe}{a data.frame containing gas measurements (see \code{gastype}
below), water vapor measurements (see \code{H2O_col} below)
and the following columns: \code{UniqueID}, \code{Etime}, and
the precision of the instrument for each gas (see description below).}
}
\value{
a plot for each incubation present in \code{dataframe} with all gas
measurements available. When CH4 measurements are detected, the plot
also highlights data points likely related to bubbling events.
}
\description{
A simple plot to provide an overview of gas measurements
}
\examples{
load(data_example_1)
plot_incubations(mydata)


}
