% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separated_flux_plot.R
\name{separated_flux_plot}
\alias{separated_flux_plot}
\title{Title}
\usage{
separated_flux_plot(
  flux.results,
  dataframe,
  gastype,
  shoulder = 30,
  plot.legend = c("SD"),
  plot.display = c("prec"),
  quality.check = TRUE,
  flux.unit = NULL,
  flux.term.unit = NULL,
  p.val.disp = "round",
  side = "left"
)
}
\arguments{
\item{flux.results}{blabla}

\item{dataframe}{blabla}

\item{gastype}{blabla}

\item{shoulder}{blabla}

\item{plot.legend}{blabla}

\item{plot.display}{blabla}

\item{quality.check}{blabla}

\item{flux.unit}{blabla}

\item{flux.term.unit}{blabla}

\item{p.val.disp}{blabla}

\item{side}{blabla}
}
\value{
blabla
}
\description{
Title
}
\examples{
blabla

}
