% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickflux.R
\name{clickflux}
\alias{clickflux}
\title{Title}
\usage{
clickflux(
  dataframe,
  myauxfile,
  shoulder,
  gastype,
  plot.lim,
  fluxSeparation,
  displayPlots
)
}
\arguments{
\item{dataframe}{blabla}

\item{myauxfile}{blabla}

\item{shoulder}{blabla}

\item{gastype}{blabla}

\item{plot.lim}{blabla}

\item{fluxSeparation}{blabla}

\item{displayPlots}{blabla}
}
\value{
blabla
}
\description{
Title
}
\examples{
blabla

}
